+++ usda-models/Weight.cs
@@ -4,8 +4,8 @@ namespace UsdaCosmos
     {
         public string FoodId { get; set; }
         public string Sequence { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string Description { get; set; }
-        public decimal WeightGrams { get; set; }
+        public double WeightGrams { get; set; }
     }
 }
