+++ usda-web-api/Models/FoodItemNutrientJson.cs
@@ -4,6 +4,6 @@ namespace UsdaCosmosJson
 {
     public class FoodItemNutrientJson : FoodItemJson 
     {
-        public decimal AmountInHundredGrams { get; set; }
+        public double AmountInHundredGrams { get; set; }
     }
 }
